refactored into python module system, WIP not tested for bugs
