+++ unified_pipeline/setup.py
@@ -1,0 +1,6 @@
+from setuptools import find_packages, setup
+
+setup(
+    name='src',
+    packages=find_packages(),
+)
